==> notebook/4_RAG_DIY.ipynb <==
# 👋👋👋自己動手做看看👋👋👋

沒有做過 RAG / embedding 的朋友可以嘗試自己動手做

---

### 資料集

`/home/jovyan/0_cold_jokes.csv`

網路上搜集的台灣的冷笑話，來源也是網路上搜集的 [https://wayne265265.pixnet.net/blog/post/216422544](https://wayne265265.pixnet.net/blog/post/216422544)

兩個欄位: question 與 anwser

### 為何要冷笑話？

1. 資料集的內容相當無用，只是讓同學試著操作 RAG
2. chatGPT 到今天還是很難產生讓人滿意的冷笑話（效能低於平均人類），透過 RAG 讓 LLM 產生本來產不好的內容：講出台灣人聽得懂的笑話
3. 未來可以依據自己需要的內容，建立有意義的知識庫。

---

### 目標：讓 openai 可以依據提供的知識（冷笑話）回答模型（gpt-4o-mini）本來不知道的問題

```
print(ask("屈原投江最後說了什麼？"))
```

我不知道。

原文依據：提供的內容中並未提及屈原投江最後說了什麼。

### 輸入冷笑話到 Qdrant 後，透過 RAG 回答

```
print(ask("屈原投江最後說了什麼？"))
```

根據提供的內容，屈原投江最後說的是：

**「偶..偶溺江（お兄ちゃん）」**

📌 **內容原文依據：**
question：屈原投江最後說了什麼？
answer：偶..偶溺江（お兄ちゃん）

---

### 提示

1. 可以複製 2-3 的 def function
2. 需要透過 qdrant.create_collection 建立新的 collection，不要跟現有的 collection 混在一起
3. 沒做出來也不會影響後面的內容

! pip install pandas

### 在沒有補笑話資料集的時後，應該答不出笑話的內容

docs = []
query = "先進船的人會先說什麼？"

docs = query_docs(
    query=query,
    collection_name="covid-qa-3-large", # 問 covid-qa collection，應該問不到笑話
    model="")

answer = generate_answer(
    query=query, 
    docs=docs, 
    model="")

print("\n🧠 回答：")
print(answer)

### 使用冷笑話資料集

import pandas as pd
joke_df = pd.read_csv('/home/jovyan/cold_jokes.csv')
joke_df.sample(n=5)

### 把 Code 寫在這裡

# 產生 embedding

# 將 embedding 寫入 Qdrant

#collection_name="取一個新的名字"

# 使用 OpenAI chat 做 RAG

### 使用冷笑話資料集應該可以回答出冷笑話問題

docs = []
query = "先進船的人會先說什麼？"

docs = query_docs(query, new_collection_name)
answer = generate_answer(query, docs)

print("\n🧠 回答：")
print(answer)
